#3370 move command inside stimmzettelerfassung
