+++ wls-gui-wahllokalsystem/src/composables/dse/stimmzettelerfassung/command/commandHandlers.ts
@@ -1,4 +1,4 @@
-import type { CommandHandler } from "@/types/dse/command/CommandHandler.ts";
+import type { CommandHandler } from "@/types/dse/stimmzettelerfassung/command/CommandHandler.ts";
 
 import { useAddVotesToSingleKandidatHandler } from "@/composables/dse/stimmzettelerfassung/command/addVotesToSingleKandidatHandler.ts";
