+++ wls-gui-wahllokalsystem/src/types/dse/stimmzettelerfassung/command/CommandHandler.ts
